# go-auth/jwt-auth/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Authentication API with JWT",
        "title": "JWT Auth API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/tasks": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Register a new task on database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Create new task",
                "parameters": [
                    {
                        "description": "Task data",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.Task"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Task created",
                        "schema": {
                            "$ref": "#/definitions/dtos.Message"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIError"
                        }
                    },
                    "409": {
                        "description": "Task already exists",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIError"
                        }
                    }
                }
            }
        },
        "/tasks/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a task by its id from database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tasks"
                ],
                "summary": "Get task by id",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"6817d32b97e0903fae78beec\"",
                        "description": "Task Id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task by id",
                        "schema": {
                            "$ref": "#/definitions/dtos.Message"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIError"
                        }
                    }
                }
            }
        },
        "/users": {
            "post": {
                "description": "Register a new user on database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.User"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User created",
                        "schema": {
                            "$ref": "#/definitions/dtos.Message"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIError"
                        }
                    },
                    "409": {
                        "description": "User already exists",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIError"
                        }
                    }
                }
            }
        },
        "/users/login": {
            "post": {
                "description": "Check if user's email and password are valid",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "User data",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User login",
                        "schema": {
                            "$ref": "#/definitions/dtos.Token"
                        }
                    },
                    "400": {
                        "description": "Erro while trying to login",
                        "schema": {
                            "$ref": "#/definitions/dtos.APIError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dtos.APIError": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "dtos.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "dtos.Task": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "This is an example task."
                },
                "title": {
                    "type": "string",
                    "example": "New Task"
                }
            }
        },
        "dtos.Token": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "dtos.User": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john.doe@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "12345a!"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Value: Bearer abc... (Bearer+space+token)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}